src/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
python-decouple>=3.8
